default to print-method for EDN encoding
This allows paradise to co-exist with existing EDN encodings via print-method. To support this, a new single-arity fn return value option was added to edn-encode and a new print-method method that calls these fns on the Writer implemenation where we're writing the EDN being generated. The README was updated with details about this and data reader interop.